rm unused code

+++ tester/api_config/config_analyzer.py
@@ -473,11 +473,11 @@ class TensorConfig:
                 if (index is not None and index == 0) or  (key is not None and key == "fpn_rois"):
                     num = self.shape[0]
                     self.numpy_tensor = numpy.zeros(self.shape)
-                    self.numpy_tensor = numpy.random.randint(1, 1024, [num, 4]).astype("float32") 
+                    self.numpy_tensor = numpy.random.randint(1, 1024, [num, 4])
                     self.numpy_tensor[:, 0] += numpy.random.random([num])
                     self.numpy_tensor[:, 1] += numpy.random.random([num])
-                    self.numpy_tensor[:, 2] = self.numpy_tensor[:, 0] + numpy.random.randint(1, 1024, [num]).astype("float32")+numpy.random.random([num])
-                    self.numpy_tensor[:, 3] = self.numpy_tensor[:, 1] + numpy.random.randint(1, 1024, [num]).astype("float32")+numpy.random.random([num])
+                    self.numpy_tensor[:, 2] = self.numpy_tensor[:, 0] + numpy.random.randint(1, 1024, [num])+numpy.random.random([num])
+                    self.numpy_tensor[:, 3] = self.numpy_tensor[:, 1] + numpy.random.randint(1, 1024, [num])+numpy.random.random([num])
                     if not hasattr(api_config, "num"):
                         api_config.num = num
                 elif (index is not None and index == 6 ) or (key is not None and key == "rois_num"):
